Set storetype for keystore
While compiling I noticed a message mentioning that the default jks is
a proprietary format, along with a recommendation to perhaps use an open
standard such as pkcs12. This sounds reasonable.

However, pkcs12 does not support different store and key passwords. Given that
the Makefile defaults to using the same pass for both it's perhaps still
reasonable to use it.